requirements_image.txt
# Image Processing (Python 3.13 compatible)
Pillow>=10.4.0
opencv-python>=4.10.0

# Face Detection
face-recognition>=1.3.0
# Note: dlib may need special handling on Mac, see below

# OCR (Text Detection)
easyocr>=1.7.0

# Object Detection
ultralytics>=8.0.0

# Deep Learning Backends
torch>=2.1.0
torchvision>=0.16.0

# Utilities
numpy>=1.24.0,<2.0.0
